+++ src/retrieval_tasks/retrieval_hybrid.py
@@ -10,13 +10,13 @@ from contextlib import contextmanager
 import numpy as np
 import pandas as pd
 from datasets import Dataset
-from rank_bm25 import BM25Plus
 from torch.nn.functional import normalize
-from transformers import AutoModel
 from typing import List, Optional, Tuple, NoReturn
 from tqdm.auto import tqdm
 
 from retrieval import Retrieval
+from retrieval_syntactic import Syntactic
+from retrieval_semantic import Semantic
 from src.utils import set_seed
 
 set_seed(2024)
@@ -28,11 +28,6 @@ def timer(name):
     t0 = time.time()
     yield
     logging.info(f"[{name}] done in {time.time() - t0:.3f} s")
-
-def mean_pooling(model_output, attention_mask):
-    token_embeddings = model_output[0] 
-    input_mask_expanded = attention_mask.unsqueeze(-1).expand(token_embeddings.size()).float()
-    return torch.sum(token_embeddings * input_mask_expanded, 1) / torch.clamp(input_mask_expanded.sum(1), min=1e-9)
 
 class HybridSearch(Retrieval):
     def __init__(
@@ -54,69 +49,34 @@ class HybridSearch(Retrieval):
         self.tokenize_fn = tokenize_fn
 
         self.dense_model_name = dense_model_name
-        self.dense_tokenize_fn = AutoTokenizer.from_pretrained(
-            self.dense_model_name
-        )
 
         self.sparse_embeder = None
-        self.dense_embeder = AutoModel.from_pretrained(
-            self.dense_model_name
-        )
+        self.dense_embeder = Semantic(self.dense_model_name)
         self.sparse_embeds = None
         self.dense_embeds = None
 
     def get_sparse_embedding(self, question=None, contexts=None):
-        vectorizer_path = os.path.join(self.data_path, "BM25Plus_sparse_vectorizer.bin")
+        vectorizer_path = os.path.join(self.data_path, "sparse_vectorizer.bin")
         if contexts is not None:
             self.contexts = contexts
-            self.sparse_embeder = BM25Plus([self.tokenize_fn(doc) for doc in self.contexts], k1=1.837782128608009, b=0.587622663072072, delta=1.1490)
+            self.sparse_embeder = Syntactic(self.tokenize_fn, contexts=self.contexts)
 
         if question is not None:
             if not hasattr(self.sparse_embeder, 'vocabulary_'):
-                vectorizer_path = os.path.join(self.data_path, "sparse_vectorizer.bin")
-                if os.path.isfile(vectorizer_path):
-                    with open(vectorizer_path, "rb") as f:
-                        self.sparse_embeder = pickle.load(f)
-                    print("Sparse vectorizer loaded for transforming the query.")
-                else:
-                    raise ValueError("The Sparse vectorizer is not fitted. Please run get_sparse_embedding() first.")
+                self.sparse_embeder = Syntactic(self.tokenize_fn, contexts=self.contexts, vectorizer_path=vectorizer_path, save_embedding=True)
             return self.sparse_embeder.transform(question)
 
         if question is None and contexts is None:
-            if os.path.isfile(vectorizer_path):
-                with open(vectorizer_path, "rb") as f:
-                    self.sparse_embeder = pickle.load(f)
-                print("Sparse vectorizer and embeddings loaded.")
-            else:
-                print("Fitting sparse vectorizer and building embeddings.")
-                self.sparse_embeder = BM25Plus([self.tokenize_fn(doc) for doc in self.contexts], k1=1.837782128608009, b=0.587622663072072, delta=1.1490)
-                with open(vectorizer_path, "wb") as f:
-                    pickle.dump(self.sparse_embeder, f)
-                print("Sparse vectorizer and embeddings saved.")
+            self.sparse_embeder = Syntactic(self.tokenize_fn, contexts=self.contexts, vectorizer_path=vectorizer_path, save_embedding=True)
 
     def get_dense_embedding(self, question=None, contexts=None, batch_size=64):
         if contexts is not None:
             self.contexts = contexts
-            self.dense_embeder.to(self.device)
-            encoded_input = self.dense_tokenize_fn(
-                self.contexts, padding=True, truncation=True, return_tensors='pt'
-            ).to(self.device)
-            with torch.no_grad():
-                model_output = self.dense_embeder(**encoded_input)
-            sentence_embeddings = mean_pooling(model_output, encoded_input['attention_mask'])
-            self.dense_embeds = sentence_embeddings.cpu()
+            self.dense_embeds = self.dense_embeder.output(self.contexts).cpu()
 
         if question is not None:
-            self.dense_embeder.to(self.device)
-            encoded_input = self.dense_tokenize_fn(
-                question, padding=True, truncation=True, return_tensors='pt'
-            ).to(self.device)
-            with torch.no_grad():
-                model_output = self.dense_embeder(**encoded_input)
-            sentence_embeddings = mean_pooling(model_output, encoded_input['attention_mask'])
+            sentence_embeddings = self.dense_embeder.output(question)
             sentence_embeddings = normalize(sentence_embeddings, p=2, dim=1)
-            self.dense_embeder.cpu()
-            del encoded_input
             return sentence_embeddings.cpu()
 
         if question is None and contexts is None:
@@ -130,16 +90,10 @@ class HybridSearch(Retrieval):
             else:
                 print("Building passage dense embeddings in batches.")
                 self.dense_embeds = torch.zeros(len(self.contexts), self.dense_embeder.config.hidden_size)
-                self.dense_embeder.to(self.device)
 
                 for i in tqdm(range(0, len(self.contexts), batch_size), desc="Encoding passages"):
                     batch_contexts = self.contexts[i:i+batch_size]
-                    encoded_input = self.dense_tokenize_fn(
-                        batch_contexts, padding=True, truncation=True, return_tensors='pt'
-                    ).to(self.device)
-                    with torch.no_grad():
-                        model_output = self.dense_embeder(**encoded_input)
-                    sentence_embeddings = mean_pooling(model_output, encoded_input['attention_mask'])
+                    sentence_embeddings = self.dense_embeder.output(batch_contexts)
                     sentence_embeddings = normalize(sentence_embeddings, p=2, dim=1)
                     self.dense_embeds[i] = sentence_embeddings.cpu()
                     del encoded_input, model_output, sentence_embeddings
